Port to AHK v2
I have ported the AHK code so its able to run on AHK v2. The code has been stable but there could be outliers or problems that i haven’t encountered yet so be warned. The two things missing from this conversion is the validation of whether a method exists. So, if the user presses a key that hasn’t been updated, it will result in an error so keep that in mind. Another missing part is the extra stuff for killing the server and ahk in the lib.ahk.

For some reason I was experiencing problems with killing the node server followed by killing the AHK for some odd reason so for now i moved it over to the main ahk to save time. If you want to tackle that feel free to do so. I just wanted to get this released as quickly as possible so others can start using AHK v2 or fix any issues that i may have missed.

In addition, I am currently working on integrating a plugin with Flow-Launcher. There is also a possibility of integrating with Powertoys Run, but for now, my focus is on Flow since it seems like a more agile system since i can build the plugin with python.

I also have plans to enhance the code options by allowing users to set parameter values for the methods. Improving the ability of the plugin on stream deck. I also have plans to add descriptions/flavour text to functions. Which should improve the aesthetics of hooking HotkeylessAHK to other things.